other: Add autocomplete file generation
