+++ Cargo.toml
@@ -10,6 +10,7 @@ categories = ["command-line-utilities", "visualization"]
 description = "A cross-platform graphical process/system monitor with a customizable interface and a multitude of features. Supports Linux, macOS, and Windows."
 readme = "README.md"
 default-run = "btm"
+build = "build.rs"
 
 [[bin]]
 name = "btm"
@@ -54,6 +55,9 @@ winapi = "0.3.9"
 [dev-dependencies]
 assert_cmd = "1.0"
 predicates = "1"
+
+[build-dependencies]
+clap = "2.33"
 
 [package.metadata.deb]
 section = "utility"
